add basic logic for save item to a cart

+++ common-dto/src/main/java/com/common/lib/cartModule/itemDto/ItemDto.java
@@ -8,7 +8,7 @@ public record ItemDto(
         Long productId,
         @NonNull
         @Positive
-        Integer quantity,
+        Short quantity,
         @NonNull
         String priceSnapshot
 ) {
